requirements.txt
uvicorn
fastapi
pymupdf
python-multipart
python-dotenv
google-generativeai
langchain
langgraph
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles